Add the ability to pick up shifts
This will allow TAs to pick up other shifts; however, since the
Instructor view are not yet implemented, there is no way for the swaps
to be approved. This means that the shifts effectively disappear once
somone picks them up (though as soon as the instructor view is
implemented, there will be a backlog of approvals to issue).